+++ src/_module.c
@@ -18,6 +18,8 @@
 #include "complement.c"
 #include "const.c"
 #include "flip.c"
+#include "packed.c"
+
 #include "nth.c"
 
 #include "argminmax.c"
@@ -153,6 +155,8 @@ PyDoc_STRVAR(PyIU_module_doc, "API: C Functions\n----------------");
         &PyIUType_Complement,
         &PyIUType_Constant,
         &PyIUType_Flip,
+        &PyIUType_Packed,
+
         &PyIUType_Nth,
 
         &PyIUType_Accumulate,