+++ src/packed.c
@@ -5,45 +5,34 @@
 typedef struct {
     PyObject_HEAD
     PyObject *func;
-    PyObject *iterator;
-} PyIUObject_Starfilter;
+} PyIUObject_Packed;
 
-PyTypeObject PyIUType_Starfilter;
+PyTypeObject PyIUType_Packed;
 
 /******************************************************************************
  * New
  *****************************************************************************/
 
 static PyObject *
-starfilter_new(PyTypeObject *type,
-               PyObject *args,
-               PyObject *kwargs)
+packed_new(PyTypeObject *type,
+           PyObject *args,
+           PyObject *kwargs)
 {
-    static char *kwlist[] = {"pred", "iterable", NULL};
-    PyIUObject_Starfilter *self;
+    PyIUObject_Packed *self;
 
-    PyObject *iterable;
-    PyObject *iterator;
     PyObject *func;
 
     /* Parse arguments */
-    if (!PyArg_ParseTupleAndKeywords(args, kwargs, "OO:starfilter", kwlist,
-                                     &func, &iterable)) {
+    if (!PyArg_UnpackTuple(args, "packed", 1, 1, &func)) {
         return NULL;
     }
 
-    /* Create and fill struct */
-    iterator = PyObject_GetIter(iterable);
-    if (iterator == NULL) {
-        return NULL;
-    }
-    self = (PyIUObject_Starfilter *)type->tp_alloc(type, 0);
+    /* Create struct */
+    self = (PyIUObject_Packed *)type->tp_alloc(type, 0);
     if (self == NULL) {
-        Py_DECREF(iterator);
         return NULL;
     }
     Py_INCREF(func);
-    self->iterator = iterator;
     self->func = func;
     return (PyObject *)self;
 }
@@ -53,10 +42,9 @@ starfilter_new(PyTypeObject *type,
  *****************************************************************************/
 
 static void
-starfilter_dealloc(PyIUObject_Starfilter *self)
+packed_dealloc(PyIUObject_Packed *self)
 {
     PyObject_GC_UnTrack(self);
-    Py_XDECREF(self->iterator);
     Py_XDECREF(self->func);
     Py_TYPE(self)->tp_free(self);
 }
@@ -66,56 +54,42 @@ starfilter_dealloc(PyIUObject_Starfilter *self)
  *****************************************************************************/
 
 static int
-starfilter_traverse(PyIUObject_Starfilter *self,
-                    visitproc visit,
-                    void *arg)
+packed_traverse(PyIUObject_Packed *self,
+                visitproc visit,
+                void *arg)
 {
-    Py_VISIT(self->iterator);
     Py_VISIT(self->func);
     return 0;
 }
 
 /******************************************************************************
- * Next
+ * Call
  *****************************************************************************/
 
 static PyObject *
-starfilter_next(PyIUObject_Starfilter *self)
+packed_call(PyIUObject_Packed *self,
+            PyObject *args,
+            PyObject *kwargs)
 {
-    PyObject *item, *val, *newargs;
-    PyObject *(*iternext)(PyObject *);
-    int ok;
-
-    iternext = *Py_TYPE(self->iterator)->tp_iternext;
-    while (item = iternext(self->iterator)) {
-        if (!PyTuple_CheckExact(item)) {
-            newargs = PySequence_Tuple(item);
-            if (newargs == NULL) {
-                Py_DECREF(item);
-                return NULL;
-            }
-        } else {
-            Py_INCREF(item);
-            newargs = item;
-        }
-        val = PyObject_Call(self->func, newargs, NULL);
-        Py_DECREF(newargs);
-        if (val == NULL) {
-            Py_DECREF(item);
-            return NULL;
-        }
-        ok = PyObject_IsTrue(val);
-        Py_DECREF(val);
-
-        if (ok > 0) {
-            return item;
-        }
-        Py_DECREF(item);
-        if (ok < 0) {
-            return NULL;
-        }
+    PyObject *packed, *newpacked, *res;
+    if (!PyArg_UnpackTuple(args, "packed.__call__", 1, 1, &packed)) {
+        return NULL;
     }
-    return NULL;
+
+    Py_INCREF(packed);
+
+    if (!PyTuple_CheckExact(packed)) {
+        newpacked = PySequence_Tuple(packed);
+        Py_DECREF(packed);
+        if (newpacked == NULL) {
+            return NULL;
+        }
+        packed = newpacked;
+    }
+
+    res = PyObject_Call(self->func, packed, NULL);
+    Py_DECREF(packed);
+    return res;
 }
 
 /******************************************************************************
@@ -123,17 +97,18 @@ starfilter_next(PyIUObject_Starfilter *self)
  *****************************************************************************/
 
 static PyObject *
-starfilter_reduce(PyIUObject_Starfilter *self)
+packed_reduce(PyIUObject_Packed *self,
+              PyObject *unused)
 {
-    return Py_BuildValue("O(OO)", Py_TYPE(self), self->func, self->iterator);
+    return Py_BuildValue("O(O)", Py_TYPE(self), self->func);
 }
 
 /******************************************************************************
  * Methods
  *****************************************************************************/
 
-static PyMethodDef starfilter_methods[] = {
-    {"__reduce__", (PyCFunction)starfilter_reduce, METH_NOARGS, PYIU_reduce_doc},
+static PyMethodDef packed_methods[] = {
+    {"__reduce__", (PyCFunction)packed_reduce, METH_NOARGS, PYIU_reduce_doc},
     {NULL, NULL}
 };
 
@@ -141,63 +116,66 @@ static PyMethodDef starfilter_methods[] = {
  * Docstring
  *****************************************************************************/
 
-PyDoc_STRVAR(starfilter_doc, "starfilter(pred, iterable)\n\
+PyDoc_STRVAR(packed_doc, "packed(func)\n\
 --\n\
 \n\
-Like :py:func:`filter` but unpacks the current item in `iterable` when \n\
-calling `pred`. This is similar to the difference between :py:func:`map` and \n\
-:py:func:`itertools.starmap`.\n\
+Class that always returns ``func(*x)`` when called with ``packed(func)(x)``.\n\
 \n\
 .. versionadded:: 0.3\n\
 \n\
 Parameters\n\
 ----------\n\
-pred : callable\n\
-    The predicate function that is called to determine if the items should\n\
-    be kept.\n\
+func : callable\n\
+    The function that should be called when the packed-instance is called.\n\
 \n\
-    .. note::\n\
-       Unlike :py:func:`filter` the `pred` cannot be ``None``.\n\
-\n\
-iterable : iterable\n\
-    `Iterable` containing the elements.\n\
-\n\
-Returns\n\
+Methods\n\
 -------\n\
-iterator : generator\n\
-    A normal iterator over `iterable` containing only the items where \n\
-    ``pred(*item)`` is ``True``.\n\
-\n\
-Notes\n\
------\n\
-This is identical to ``filter(lambda x: pred(*x), iterable)`` but faster.\n\
+__call__(x)\n\
+    Returns `func(\\*x)`.\n\
 \n\
 Examples\n\
 --------\n\
-A simple example::\n\
+Creating `packed` instances::\n\
+\n\
+    >>> from iteration_utilities import packed\n\
+    >>> from operator import eq\n\
+    >>> five = packed(eq)\n\
+    >>> five((2, 2))\n\
+    True\n\
+\n\
+This is a convenience class that emulates the behaviour of \n\
+:py:func:`itertools.starmap` (compared to :py:func:`map`)::\n\
+\n\
+    >>> from itertools import starmap\n\
+    >>> list(map(packed(eq), [(2, 2), (3, 3), (2, 3)]))\n\
+    [True, True, False]\n\
+    >>> list(starmap(eq, [(2, 2), (3, 3), (2, 3)]))\n\
+    [True, True, False]\n\
+\n\
+and :py:func:`iteration_utilities.starfilter` (compared to :py:func:`filter`)::\n\
 \n\
     >>> from iteration_utilities import starfilter\n\
-    >>> from operator import eq\n\
-    >>> list(starfilter(eq, zip([1,2,3], [2,2,2])))\n\
-    [(2, 2)]\n\
+    >>> list(filter(packed(eq), [(2, 2), (3, 3), (2, 3)]))\n\
+    [(2, 2), (3, 3)]\n\
+    >>> list(starfilter(eq, [(2, 2), (3, 3), (2, 3)]))\n\
+    [(2, 2), (3, 3)]\n\
 \n\
-See also\n\
---------\n\
-filter\n\
-iteration_utilities.packed\n\
+Of course in these cases the appropriate `star`-function can be used but \n\
+in case a function does not have such a convenience function already `packed`\n\
+can be used.\n\
 ");
 
 /******************************************************************************
  * Type
  *****************************************************************************/
 
-PyTypeObject PyIUType_Starfilter = {
+PyTypeObject PyIUType_Packed = {
     PyVarObject_HEAD_INIT(NULL, 0)
-    "iteration_utilities.starfilter",                   /* tp_name */
-    sizeof(PyIUObject_Starfilter),                      /* tp_basicsize */
+    "iteration_utilities.packed",                       /* tp_name */
+    sizeof(PyIUObject_Packed),                          /* tp_basicsize */
     0,                                                  /* tp_itemsize */
     /* methods */
-    (destructor)starfilter_dealloc,                     /* tp_dealloc */
+    (destructor)packed_dealloc,                         /* tp_dealloc */
     0,                                                  /* tp_print */
     0,                                                  /* tp_getattr */
     0,                                                  /* tp_setattr */
@@ -207,21 +185,21 @@ PyTypeObject PyIUType_Starfilter = {
     0,                                                  /* tp_as_sequence */
     0,                                                  /* tp_as_mapping */
     0,                                                  /* tp_hash */
-    0,                                                  /* tp_call */
+    (ternaryfunc)packed_call,                           /* tp_call */
     0,                                                  /* tp_str */
     PyObject_GenericGetAttr,                            /* tp_getattro */
     0,                                                  /* tp_setattro */
     0,                                                  /* tp_as_buffer */
     Py_TPFLAGS_DEFAULT | Py_TPFLAGS_HAVE_GC |
         Py_TPFLAGS_BASETYPE,                            /* tp_flags */
-    starfilter_doc,                                     /* tp_doc */
-    (traverseproc)starfilter_traverse,                  /* tp_traverse */
+    packed_doc,                                         /* tp_doc */
+    (traverseproc)packed_traverse,                      /* tp_traverse */
     0,                                                  /* tp_clear */
     0,                                                  /* tp_richcompare */
     0,                                                  /* tp_weaklistoffset */
-    PyObject_SelfIter,                                  /* tp_iter */
-    (iternextfunc)starfilter_next,                      /* tp_iternext */
-    starfilter_methods,                                 /* tp_methods */
+    0,                                                  /* tp_iter */
+    0,                                                  /* tp_iternext */
+    packed_methods,                                     /* tp_methods */
     0,                                                  /* tp_members */
     0,                                                  /* tp_getset */
     0,                                                  /* tp_base */
@@ -230,7 +208,7 @@ PyTypeObject PyIUType_Starfilter = {
     0,                                                  /* tp_descr_set */
     0,                                                  /* tp_dictoffset */
     0,                                                  /* tp_init */
-    PyType_GenericAlloc,                                /* tp_alloc */
-    starfilter_new,                                     /* tp_new */
+    0,                                                  /* tp_alloc */
+    packed_new,                                         /* tp_new */
     PyObject_GC_Del,                                    /* tp_free */
 };